aula-3

+++ server/src/routes/memories.ts
@@ -3,8 +3,14 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -28,6 +34,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+    if (!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
     return memory;
   });
 
@@ -49,7 +58,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -63,8 +72,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
 
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
 
-    const memory = await prisma.memory.update({
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -78,16 +96,25 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories:id", async (request) => {
+  app.delete("/memories:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
     const { id } = paramsSchema.parse(request.params);
-    const memory = await prisma.memory.delete({
+
+    let memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
-    return memory;
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    });
   });
 }
